backend/pyproject.toml
[project]
name = "douchebags-open-backend"
version = "0.1.0"
description = "Backend API for The Douchebags Open MVP"
requires-python = ">=3.12"
dependencies = [
  "alembic>=1.13,<2.0",
  "email-validator>=2.2,<3.0",
  "fastapi>=0.115,<1.0",
  "passlib[argon2]>=1.7,<2.0",
  "pillow>=10.4,<12.0",
  "psycopg[binary]>=3.2,<4.0",
  "pydantic-settings>=2.4,<3.0",
  "python-jose[cryptography]>=3.3,<4.0",
  "python-multipart>=0.0.9,<1.0",
  "sqlalchemy>=2.0,<3.0",
  "uvicorn[standard]>=0.30,<1.0",
]

[project.optional-dependencies]
dev = [
  "httpx>=0.27,<1.0",
  "pytest>=8.3,<9.0",
]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]

